Convert Cristin monthly data to JSON format

This function converts the provided Cristin monthly data into JSON format, which can be used for various purposes.

CristinJson(
  monthlies,
  unit.paths,
  sdg.data = NULL,
  local.storage = NULL,
  user.cards = TRUE,
  use.multisession = FALSE,
  min.multisession = 25,
  n.workers = NULL,
  handler = "cli",
  restore.defaults = FALSE,
  full.update = FALSE,
  lang = "nn",
  silent = FALSE,
  log = list()
)

Arguments

monthlies

Data containing monthly information from Cristin.

unit.paths

Data containing the Zotero paths of Cristin units.

sdg.data

Data containing information about Sustainable Development Goals (SDGs). Default is NULL.

local.storage

Path to local storage directory. Default is NULL.

user.cards

Logical, indicating whether to update user information from INN. Default is TRUE.

use.multisession

Logical. If TRUE (default), parallel processing using multisession is employed; otherwise, processing is sequential.

min.multisession

Minimum number of results for using multisession. Default: 25

n.workers

Optional integer for the number of workers to be used in multisession mode. If NULL, it defaults to the number of available cores minus one (with a minimum of one).

handler

The progress handler to be used by the progressr package. If NULL and silent is FALSE, it defaults to "txtprogressbar". When silent is TRUE, the handler is set to "void".

restore.defaults

Logical. If TRUE (default), the current future plan is saved and restored upon exit.

full.update

Logical, indicating whether to perform a full update of user information from INN. Default is FALSE.

lang

Language for localization ("nb", "nn", "no", or "en"). Defaults to "nn".

silent

Logical, indicating whether to suppress log messages. Default is FALSE.

log

List to store log messages. Default is an empty list.

Value

A JSON-formatted string representing the converted Cristin monthly data.

Details

Used with `CristinMonthly` to create month-to-month bibliography in JSON format

Examples

# \donttest{
  # Create monthlies for unit 209.5.10.0
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "Io1bwAiOjB1jPgUNegjJhJxF",
      user = FALSE
    ),
    unit.id = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = TRUE,
    use.citeproc = FALSE
  ) |>
    c2z:::GoFish()

  # Create json data
  example.json <- CristinJson(
    example$monthlies,
    example$unit.paths,
    user.cards = FALSE
  ) |>
    jsonlite::prettify() |>
    c2z:::GoFish()

  if (any(!is.na(example.json))) cat(example.json)
# }