create a tibble with information about selected Cristin units

Create a tibble with information about (nested) units in Cristin (e.g., A University -> Faculties -> Departments -> Groups). The tibble can than be used to extract data for each unit from Cristin. Used by `CristinMonthly`

CristinUnits(
  unit.id,
  subunits = TRUE,
  recursive = FALSE,
  ancestors = FALSE,
  lang = "en"
)

Arguments

unit.id

Unit to search for

subunits

Add subunits, Default: TRUE

recursive

Search for (nested) sub-units, Default: FALSE

ancestors

Search for parent units, Default: FALSE

lang

PARAM_DESCRIPTION, Default: `en`

Value

A tibble containing information about selected units

Details

Used with `CristinMontlhy` to create month-to-month bibliography of selected units

See also

Examples

# \donttest{
# Find units for Inland University
  CristinUnits("209.0.0.0") |>
    dplyr::select(id, name) |>
    print(width = 80)
#> # A tibble: 8 × 2
#>   id        name                                                               
#>   <chr>     <chr>                                                              
#> 1 209.0.0.0 University of Inland Norway                                        
#> 2 209.7.0.0 Administration                                                     
#> 3 209.2.0.0 Faculty of Applied Ecology, Agricultural Sciences and Biotechnology
#> 4 209.3.0.0 Faculty of Audiovisual Media and Creative Technologies             
#> 5 209.5.0.0 Faculty of Education                                               
#> 6 209.4.0.0 Faculty of Social and Health Sciences                              
#> 7 209.6.0.0 Inland School of Business and Social Sciences                      
#> 8 209.1.0.0 The Norwegian Film School                                          
# }