Create a month-to-month bibliography of selected units

Create a bibliography in a newsletter format from month-to-month From either a specified unit or a set of units (e.g., A University -> Faculties -> Departments -> Groups).

CristinMonthly(
  zotero,
  unit.key,
  unit.recursive = TRUE,
  sdg.script = NULL,
  sdg.model = NULL,
  sdg.cutoff = 0.98,
  sdg.host = NULL,
  get.unpaywall = FALSE,
  get.ezproxy = FALSE,
  ezproxy.host = "inn.no",
  local.storage = NULL,
  style = "apa-single-spaced",
  locale = "nn-NO",
  start.date = Sys.Date(),
  end.date = NULL,
  use.filter = TRUE,
  filter = NULL,
  nvi = TRUE,
  check.items = TRUE,
  use.identifiers = TRUE,
  full.update = FALSE,
  lang = "nn",
  post.lang = "nn",
  post = FALSE,
  silent = FALSE,
  log = list()
)

Arguments

zotero

What Zotero library to use

unit.key

What unit to search for

unit.recursive

Find subunits of defined unit key, Default: TRUE

sdg.script

script to run SDG predictions, Default: NULL

sdg.model

model to use for SDG predictions, Default: NULL

sdg.cutoff

cut off value for SDG predictions, Default: '0.98'

sdg.host

host conducting SDG predictions, Default: NULL

get.unpaywall

Find Unpaywall resources, Default: FALSE

get.ezproxy

Use ezproxy, Default: FALSE

ezproxy.host

ezproxy host, Default: 'inn.no'

local.storage

Path to local storage of collections, items and bibliography, Default: 'NULL'

style

Citation style to use for appended bibliography and/or citations, Default: 'apa-single-spaced'

locale

Desired language format of bibliography, Default: 'nn-NO'

start.date

Results created from specified date (YYYY-MM), Default: Sys.Date()

end.date

Results created before specified date (YYYY-MM), Default: NULL

use.filter

Filter out specific items (otherwise supported by `CristinWrangler`), Default: TRUE

filter

Filter out specific item types, will default to item types usually associated with NVI, Default: NULL

nvi

Filter out Cristin items not 1/2 in NVI, Default: TRUE

check.items

Examine items for duplicates and multidepartemental publications, Default: TRUE

use.identifiers

Use if ISBN/DOI identifiers if enabled, Default: TRUE

full.update

Update bibliography for all items in library, Default: FALSE

lang

Define bibliography language (nn, nb, en), Default: 'nn'

post.lang

Define language for Zotero collections (nn, nb, en), Default: 'nn'

post

Post new items to specified Zotero library, Default: FALSE

silent

c2z is noisy, tell it to be quiet, Default: FALSE

log

A list for storing log elements, Default: list()

Value

A list with bibliography and information for defined units

Details

Used with `CristinUnits` to create month-to-month bibliography of selected units

See also

Examples

# \donttest{
  # Create monthlies for unit 209.5.10.0
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "RqlAmlH5l1KPghfCseAq1sQ1",
      user = FALSE
    ),
    unit.key = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = TRUE
  )
#> Error in c2z::Zotero(zotero = zotero, collection.key = core.location,     search.collections = FALSE, item.type = "-attachment || note",     library = TRUE, silent = silent, force = TRUE, log = zotero$log): unused argument (search.collections = FALSE)

  # Print the three first individual pages (if any)
  if (!is.null(example$monthlies)) {
    example$monthlies |>
      dplyr::select(title) |>
      print(n = 3, width = 80)
  }
#> Error in example$monthlies: object of type 'closure' is not subsettable
# }